+++ src/libs/satellite/src/storage/interfaces.rs
@@ -1,0 +1,86 @@
+use crate::get_content_chunks_store;
+use crate::hooks::invoke_assert_upload_asset;
+use crate::storage::state::{get_asset, get_config, get_rule, insert_asset, insert_asset_encoding};
+use crate::storage::store::get_public_asset_store;
+use candid::Principal;
+use junobuild_collections::types::rules::{Memory, Rule};
+use junobuild_shared::types::core::{Blob, CollectionKey};
+use junobuild_storage::interfaces::{AssertOperations, ContentStore, InsertOperations};
+use junobuild_storage::types::config::StorageConfig;
+use junobuild_storage::types::state::FullPath;
+use junobuild_storage::types::store::{Asset, AssetAssertUpload, AssetEncoding};
+
+pub struct SatelliteAssertOps;
+
+impl AssertOperations for SatelliteAssertOps {
+    fn invoke_assert_upload_asset(
+        &self,
+        caller: &Principal,
+        asset: &AssetAssertUpload,
+    ) -> Result<(), String> {
+        invoke_assert_upload_asset(caller, asset)
+    }
+}
+
+pub struct SatelliteInsertOps;
+
+impl InsertOperations for SatelliteInsertOps {
+    fn insert_state_asset_encoding(
+        &self,
+        full_path: &FullPath,
+        encoding_type: &str,
+        encoding: &AssetEncoding,
+        asset: &mut Asset,
+        rule: &Rule,
+    ) {
+        insert_asset_encoding(full_path, encoding_type, encoding, asset, rule);
+    }
+
+    fn insert_state_asset(
+        &self,
+        collection: &String,
+        full_path: &String,
+        asset: &Asset,
+        rule: &Rule,
+    ) {
+        insert_asset(collection, full_path, asset, rule);
+    }
+}
+
+pub struct SatelliteContentStore;
+
+impl ContentStore for SatelliteContentStore {
+    fn get_content_chunks(
+        &self,
+        encoding: &AssetEncoding,
+        chunk_index: usize,
+        memory: &Memory,
+    ) -> Option<Blob> {
+        get_content_chunks_store(encoding, chunk_index, memory)
+    }
+
+    fn get_public_asset(
+        &self,
+        full_path: FullPath,
+        token: Option<String>,
+    ) -> Option<(Asset, Memory)> {
+        get_public_asset_store(full_path, token)
+    }
+
+    fn get_asset(
+        &self,
+        collection: &CollectionKey,
+        full_path: &FullPath,
+        rule: &Rule,
+    ) -> Option<Asset> {
+        get_asset(collection, full_path, rule)
+    }
+
+    fn get_rule(&self, collection: &CollectionKey) -> Result<Rule, String> {
+        get_rule(collection)
+    }
+
+    fn get_config(&self) -> StorageConfig {
+        get_config()
+    }
+}
